chore: moving push project to next day into project details component
