+++ components/projects/ProjectAccordionItem.tsx
@@ -1,43 +1,26 @@
 import { useAccountsContext } from "@/api/ContextAccounts";
-import { Project, useProjectsContext } from "@/api/ContextProjects";
+import { Project } from "@/api/ContextProjects";
 import { calcRevenueTwoYears, make2YearsRevenueText } from "@/helpers/projects";
-import { addDays, format } from "date-fns";
+import { format } from "date-fns";
 import { flow, map, sum } from "lodash/fp";
-import { ArrowRightCircle, Loader2 } from "lucide-react";
-import { FC, useState } from "react";
+import { FC } from "react";
 import DefaultAccordionItem from "../ui-elements/accordion/DefaultAccordionItem";
 import ProjectDetails from "../ui-elements/project-details/project-details";
-import { Button } from "../ui/button";
 
 type ProjectAccordionItemProps = {
   project?: Project;
   showNotes?: boolean;
   onDelete?: () => void;
   disabled?: boolean;
-  allowPushToNextDay?: boolean;
 };
 
 const ProjectAccordionItem: FC<ProjectAccordionItemProps> = ({
   project,
   onDelete,
   disabled,
-  allowPushToNextDay,
   showNotes = true,
 }) => {
-  const [pushingInProgress, setPushingInProgress] = useState(false);
-  const { saveProjectDates } = useProjectsContext();
   const { getAccountNamesByIds } = useAccountsContext();
-
-  const handlePushToNextDay = async () => {
-    if (!project) return;
-    setPushingInProgress(true);
-    const result = await saveProjectDates({
-      projectId: project.id,
-      onHoldTill: addDays(new Date(), 1),
-    });
-    if (result) return;
-    setPushingInProgress(false);
-  };
 
   return (
     project && (
@@ -66,23 +49,6 @@ const ProjectAccordionItem: FC<ProjectAccordionItemProps> = ({
         ]}
         disabled={disabled}
       >
-        {allowPushToNextDay && (
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={handlePushToNextDay}
-            disabled={pushingInProgress}
-          >
-            {!pushingInProgress && (
-              <ArrowRightCircle className="w-4 h-4 mr-1" />
-            )}
-            {pushingInProgress && (
-              <Loader2 className="w-4 h-4 mr-1 animate-spin" />
-            )}
-            Push to next day
-          </Button>
-        )}
-
         <ProjectDetails
           projectId={project.id}
           showCrmDetails
